requirements.txt
langchain
faiss-cpu
transformers
torch
sentence-transformers
pymupdf
gradio
pyyaml
numpy
sentencepiece